Create a dedicated function for the broadcasting of messages to clients. Messages are of two flavors (controled by the message type argument) and are only sent to connected clients, chosen after the list of serials given in arguments, if any.
Use the new message broadcast function in autodealCheck. 

Modify the autodeal function to handle the case of tournament breaks in a special
way. If autodeal is attempted while the tournament is not running and that the tournament
is waiting for some tables to finish their hand, notify all the players.


git-svn-id: http://svn.gna.org/svn/pokersource/trunk@2787 c4f34931-41c7-4605-b615-9bb9f479c57e