Move the clock control code to testclock.py file.
Adjust the timeout of some tests. This is needed because the
artificial clocks goes faster than the real clock. Controling the
clock is not an exact science because 1) threads are used for
pokerlock and when the threads wake up they change the number of loops
of the twisted reactor, 2) the number of loop depends on the precise
twisted implementation and would therefore not be exactly the same
with twisted 1.3 or twisted 2.4. However, once the number of iteration
has been bounded for the existing twisted implementations, there is
very little chance that it will greatly change.


git-svn-id: http://svn.gna.org/svn/pokersource/trunk@2789 c4f34931-41c7-4605-b615-9bb9f479c57e