Merged table-picker-bug-13789 into trunk.  TablePicker() now returns fullest table w/ seat that matches given criteria, which closed bug#13789.
This was implemented by simply adding an ORDER BY on listTables() query to
force the tables with the most players to be seen first by the selection
function, getTableBestByCriteria() which already returns when it finds the
first table that isn't full and the player has enough money to buy in for.

Loic raised concern about doing an ORDER BY serial as well, since it slows
the query down a bit.  I left it in as it allows the semantic to be
clearly defined when there are ties.

git-svn-id: http://svn.gna.org/svn/pokersource/trunk@5982 c4f34931-41c7-4605-b615-9bb9f479c57e